Moved/edited deviceInfo code and suggested improvements
Moved most of the code from skeleton engine here, but made it look a
little neater and added some tips on how to improve when I have access
to my laptop instead of my phone.
Also added tips for custom a device info query on multiple systems.
This class is likely going to see some big changes in the future!