Fixed bugs; more overridable log() function
Added wait_For_Threads() function to wait until all threads are unregistered. There is currently a problem with this where, if a new thread is just made but is not yet registered, it may be ignored, and then the program closes before the thread can run. This is leading to the logging function wiping the log file when the program exits.
log() function gathers time and thread information before logging. A thread then opens for actually writing the message.