Added cleanup(); Fixed race condition
Threads now asynchronously add to a thread count before opening, so there is no more race condition when waiting for threads to close, even if the threads have not all been joined or are detached. There seems to be a small chance that it still happens, but logically that should not be the case and is only due to the way it compiles. Maybe. I don't know.
My thought is, even if it is a detached thread creating a new thread, as long as that first thread is alive, still counts towards the number of threads, and will increase that number before opening the new thread and killing itself, meaning it must have sustained the thread count down the line, even if the thread isn't registered yet. So I have no idea why it sometimes misses the final log.