Fixed vtable cannot be found error
It wasn't anything to do with the virtual functions. I tried making a static function and even that was not found. Turns out, I was just compiling the source files of each code entry point, not the files they included, which means that client_state_loading.cpp didn't exist to the linker since it was never compiled. I fixed that and the error went away.