GraphicsDevice work
Now compiles and is mostly done to how I'd like it. Lot of work... Not tested yet since I need to have a surface to check suitability, which requires the Window class to be completed. Soon though! Then I can move on to data serialization like I should be according to the Trello.