Made logging synchronous
The asynchronous is likely not helping much at all due to how long it takes to create threads possibly taking longer than it does to create files. It also sometimes puts the entries in out of order. Making a thread with a queue would be a pain in the butt and I don't want to have a wild thread doing nothing most of the time for many reasons.