Fixed memory leak bug
The issue was that the shader code was getting loaded into a local auto variable, which was edited and used for pointers in the create info, but then goes out of scope, which is undefined behavior. Game compiles and window opens with a working graphics pipeline. Up next will be framebuffers and command buffers, which will allow me to set the window to gray!