Work on new stack trace system
Not using default GCC stack trace since the entire GCC library has to be rebuilt, which is a pain in the butt. It may be more beneficial to make my own stack trace class anyway to have more customizability with line numbers, etc. Only problem is making it portable. The Linux/Mac portability update will be so fun, I'm sure! And the iOS/Android/ARM ones? Oh boy!

Still needs work to get/demangle the symbols, but the stack trace is kept in a vector of void* to all the function addresses.